Fix implict types

+++ contracts/ExpandedNFT.sol
@@ -426,7 +426,7 @@ contract ExpandedNFT is
         internal returns (uint256)
     {
         if (_randomMint) {
-            uint256 random = uint(keccak256(abi.encodePacked(msg.sender,block.prevrandao,gasleft()))) % dropSize;
+            uint256 random = uint256(keccak256(abi.encodePacked(msg.sender,block.prevrandao,gasleft()))) % dropSize;
             uint256 randomIndex = 1 + random;
 
             while (_perTokenMetadata[randomIndex].state != ExpandedNFTStates.UNMINTED) {
@@ -840,8 +840,8 @@ contract ExpandedNFT is
 
         require(_mintedMetadataUrl.length == count, "Data size mismatch");
 
-        for (uint i = 0; i < count; i++) {
-            uint index =  startIndex + i;
+        for (uint256 i = 0; i < count; i++) {
+            uint256 index =  startIndex + i;
             
             _perTokenMetadata[index].mintedMetadataUrl =_mintedMetadataUrl[i];
